More Updates and Improvements - STILL A WIP
Renamed scripts to not have extensions and to have similar naming
convensions. Updated docs to reflect those.

Added a time parameter to rsync in cloudkey_sync to only pull the last N
days of files to allow cleanup on the destination server as files are
processed.

Moved all files to /usr/local/bin because that's where they should be,
and updated all references in docs and scripts to point to the proper
location.

Added unifi_protect_cameras and unifi-protect.conf for querying the API
to identify cameras by their MAC addresses.