Branching out literally and metaphorically
The data this is handling is too high risk to potentially lose so I am
rewriting this with some new requirements in mind. I had a bug that
caused my prepared .TXT files to overwrite the UBV files, which
fortunately didn't lead to data loss but I need to fix it.

I'll be opening issues to track the pending changes for this.