Added Batch Function
Still a lot to do but it's coming together. _remux() will wrap around
the subprocess utility to remux files and return the mp4s since we can
have multiple outputs per remux, such as when a camera loses
connectivity.

remux_ubv_by_date() is there to help catch up but I think there's a way
to remux files as they arrive on the NAS from the sync, I just need to
always work on N-1 as they are synchronized.

Tests are lacking but they always are. Will pick those up next.