Develop -> Main (#2)
* Branching out literally and metaphorically

The data this is handling is too high risk to potentially lose so I am
rewriting this with some new requirements in mind. I had a bug that
caused my prepared .TXT files to overwrite the UBV files, which
fortunately didn't lead to data loss but I need to fix it.

I'll be opening issues to track the pending changes for this.

* Cleaned up

In preparation for a rewrite I've cleaned up this repository.

* Cloudkey Script Written

* cloudkey_rsync Permission Changes

Modified the script to set permissions after completing. This allows for
passing the -a param but overriding the destination permissions after
synchronization.

* Rewritten with Tests

Base functionality is there now for the identification of UBV files,
preparation, remux, and moving into storage.

* Added Batch Function

Still a lot to do but it's coming together. _remux() will wrap around
the subprocess utility to remux files and return the mp4s since we can
have multiple outputs per remux, such as when a camera loses
connectivity.

remux_ubv_by_date() is there to help catch up but I think there's a way
to remux files as they arrive on the NAS from the sync, I just need to
always work on N-1 as they are synchronized.

Tests are lacking but they always are. Will pick those up next.

* remux.py Mostly Working

The script will now accept arguments on the command line to allow for
cron scheduling. There's still some tweaks needed such as cleaning up
the logging language and adding in cleanup functionality, but that will
be done at a later date.

If needed someone can write a find + rm script to purge older files once
they have been processed.

* Archive Commit